move graph class to lib
