--- additional_tasks/graph/README.md
@@ -1,5 +1,0 @@
-# Класс `Graph` от Криворучко Дмитрия Игоревича
-
-Ненормальный класс `Graph`, содержащий класс `Edge` в качестве приватного поля, но не имеющий отдельного класса Vertice (допускается использование только `std::string` и целочисленных типов (`int`, `short`, `std::size_t`) в качестве названия вершин, подставляемых шаблонами) с восемью статическими методами — псевдоконструкторами, различными вспомогательными методами для удаления и поиска ребра или вершины.
-
-Может использоваться для решения специфических задач, где граф надо хранить не как список смежности, а как списки вершин и рёбер.
