--- additional_tasks/graph/src/utils.cpp
@@ -1,1 +1,0 @@
-#include "utils.hpp"
