+++ lib/src/lib/utils.cpp
@@ -1,4 +1,4 @@
-#include "util.hpp"
+#include "utils.hpp"
 
 std::string ErasedZerosStr(float number) {
   std::string str = std::to_string(number);
